requirements.txt
amqp==2.1.4
appdirs==1.4.3
appnope==0.1.0
aspy.yaml==0.3.0
bandit==1.4.0
billiard==3.5.0.2
boltons==16.1.1
brotlipy==0.6.0
cached-property==1.3.0
celery==4.0.2
certifi==2017.4.17
cffi==1.10.0
click==6.7
coverage==4.4.1
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.1
django-log-request-id==1.3.2
django-model-utils==3.0.0
django-naomi==0.8
django-webpack-loader==0.5.0
django.js-vinta==0.8.2.dev2
djangorestframework==3.6.3
fixmydjango==0.2
flake8==2.6.2
flake8-putty==0.4.0
future==0.16.0
gitdb2==2.0.0
GitPython==2.1.3
gunicorn==19.7.1
ipdb==0.10.3
ipython==6.0.0
ipython-genutils==0.2.0
isort==4.2.5
jedi==0.10.2
kombu==4.0.2
mccabe==0.5.3
mock==2.0.0
model-mommy==1.3.2
nodeenv==1.1.2
nose==1.3.7
oauthlib==2.0.2
opbeat==3.5.2
packaging==16.8
pbr==3.0.1
pexpect==4.2.1
pickleshare==0.7.4
pre-commit==0.14.0
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycodestyle==2.0.0
pycparser==2.17
pyflakes==1.2.3
Pygments==2.2.0
PyJWT==1.5.0
pyparsing==2.2.0
python-decouple==3.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.14.2
requests-oauthlib==0.8.0
safety==1.4.0
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.1
social-auth-app-django==1.2.0
social-auth-core==1.3.0
stevedore==1.21.0
tblib==1.3.2
termcolor==1.1.0
traitlets==4.3.2
tweepy==3.5.0
urllib3==1.21.1
vine==1.1.3
virtualenv==15.1.0
wcwidth==0.1.7
wheel==0.24.0
whitenoise==3.3.0
